requirements.txt
# Django framework + REST framework
django==2.0
djangorestframework==3.7.3
django-cors-headers==2.1.0
django-cleanup==1.0.1
pyjwt==1.5.3
pytz==2017.3
pillow==4.3.0

# Database adapters
psycopg2==2.7.3.2
#mysqlclient==1.3.12

# Channels framework
asgi_redis==1.4.3
channels==1.1.8
daphne==1.3.0

# Celery: Distributed Task Queue
celery==4.1.0
sqlalchemy==1.1.15

# HTTP requests
requests==2.18.4
gunicorn==19.7.1